develop/python/bot/bot-discord/bot.py
import os
import asyncio
import discord
from discord.ext import commands
from dotenv import load_dotenv

load_dotenv()

DISCORD_TOKEN = os.getenv("DISCORD_TOKEN")

class Bot(commands.Bot):
    def __init__(self):
        intents = discord.Intents.default()
        super().__init__(command_prefix="!", intents=intents)

    async def setup_hook(self):
        # Load all cogs from cogs directory
        for filename in os.listdir("./cogs"):
            if filename.endswith(".py") and not filename.startswith("_"):
                await self.load_extension(f"cogs.{filename[:-3]}")
                print(f"Loaded cog: {filename[:-3]}")
        
        # Sync commands
        await self.tree.sync()
        print("Commands synced")

    async def on_ready(self):
        print(f"Bot is ready: {self.user}")

bot = Bot()
bot.run(DISCORD_TOKEN)
