develop/python/bot/bot-discord/requirements.txt
discord.py==2.3.2
asyncssh==2.14.2
python-dotenv==1.0.1
